backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
sentence-transformers
pydantic-settings
praw
tweepy
snscrape
apscheduler
scikit-learn
numpy
pandas
torch
transformers
passlib[bcrypt]
python-jose[cryptography]
python-multipart
asyncio-mqtt
websockets
sklearn
umap-learn
hdbscan
